Add files via upload
In genome-wide association studies (GWAS), a wide variety of analysis tools were designed leading to many types of data formats. We may desire to import a dataset in non-PLINK format into PLINK to use its powerful analysis performance, or employ a dataset in PLINK formats to other analysis tools. For this purpose, we developed a tool called coPLINK, a complementary tool to PLINK, to cooperate with PLINK to implement the conversions of GWAS data formats and, to provide some additional functions such as data files comparison. The tool can implement the conversions each other not only between an existing data format and PLINK PED/BED, but also between a user-defined data format and PLINK PED. The usage and performance of the tool are similar to PLINK. The characteristics of the conversions of existing data formats and user-defined formats make it be a good assistant to PLINK or other tools and, have good potential in GWAS studies or other works.